discord-social-bot/requirements.txt
discord.py
requests
python-dotenv
apscheduler
sqlalchemy
cryptography
aiohttp